+++ src/DoctorModel.java
@@ -1,4 +1,7 @@
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.HashMap;
 
 
 /**handles all calculations, database queries, and the overall work needed to be done for handling this associated role
@@ -11,11 +14,14 @@ public class DoctorModel extends UserSuperClass {
 	private ArrayList<String> assignedNurseUsernames = new ArrayList<String>(5);
 	private ArrayList<String> scheduledPatientsUsernames = new ArrayList<String>(5);
 	// private Schedule schedule;
-	
+	private LocalDateTime[] availability;
+	private HashMap<String, ArrayList<LocalDateTime>> appointments;
 
 	
 	//constructor
-	public DoctorModel(String username, char[]password, String name, String department, String[] nurses, String[] assigPats) {
+	// public DoctorModel(String username, char[]password, String name, String department, String[] nurses, String[] assigPats, ArrayList<String> available) 
+	public DoctorModel(String username, char[]password, String name, String department, String[] nurses, String[] assigPats) 
+	{
 		
 			super(name, username, password);
 			this.department = department;
@@ -32,6 +38,7 @@ public class DoctorModel extends UserSuperClass {
 			{
 				this.scheduledPatientsUsernames.add(i);
 			}
+			// this.availability = arrayToLDTArray(available);
 			
 	}
 	
@@ -41,7 +48,8 @@ public class DoctorModel extends UserSuperClass {
 		// setScheduledPatients(ps);
 		this.scheduledPatientsUsernames.add(p);
 	}
-	
+
+
 	
 	
 	
@@ -55,6 +63,22 @@ public class DoctorModel extends UserSuperClass {
 	// public Schedule getSchedule() {
 	// 	return schedule;
 	// }
+
+	public void setAppointments(HashMap<String, ArrayList<LocalDateTime>> xs) {
+		this.appointments = xs;
+	}
+
+	public HashMap<String, ArrayList<LocalDateTime>> getAppointments() {
+		return appointments;
+	}
+
+	public void setAvailability(LocalDateTime[] xs) {
+		this.availability = xs;
+	}
+
+	public LocalDateTime[] getAvailability() {
+		return availability;
+	}
 
 	public String getDepartment() {
 		return department;